4 прака

+++ prac4/src/api/planet/delete.php
@@ -1,30 +1,28 @@
 <?php
 
-// HTTP-заголовки
-header("Access-Control-Allow-Origin: *");
 header("Content-Type: application/json; charset=UTF-8");
-header("Access-Control-Allow-Methods: DELETE");
-header("Access-Control-Max-Age: 3600");
-header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
 
 include_once "../config/database.php";
 include_once "../objects/planet.php";
 
-$database = new Database();
-$db = $database->getConnection();
+if ($_SERVER['REQUEST_METHOD'] == "DELETE") {
+    $database = new Database();
+    $db = $database->getConnection();
 
-$planet = new Planet($db);
+    $planet = new Planet($db);
 
-$data = json_decode(file_get_contents("php://input"));
+    $planet->id = $_GET["id"] ?? die();;
 
-$planet->id = $data->id;
+    if ($planet->delete()) {
+        http_response_code(200);
 
-if ($planet->delete()) {
-    http_response_code(200);
+        echo json_encode(array("message" => "Планета была удалена"), JSON_UNESCAPED_UNICODE);
+    } else {
+        http_response_code(503);
 
-    echo json_encode(array("message" => "Планета была удалена"), JSON_UNESCAPED_UNICODE);
+        echo json_encode(array("message" => "Не удалось удалить планету"));
+    }
 } else {
-    http_response_code(503);
-
-    echo json_encode(array("message" => "Не удалось удалить планету"));
+    http_response_code(400);
+    echo "Wrong Request method";
 }
